fix check board web server is up method
